Added libc namespaces and re-wrote printf.
The libc (should probably be called libk) originally had no namespaces
and polluted the global namespace. This caused issues with testing and
is just bad practice. The reason it was like this was because I copied
a lot of this code from osdev.org, which assumes a kernel written in C.
So, I finally got around to adding namespaces. One thought is to wrap all of
libc in a `libc` namespace but we'll see if that becomes an issue. Also,
the arch kernel libraries should have namespaces as well.

In addition, printf was re-written because I didn't know why it had been
written the previous way and the code was ugly. It's still a little ugly
but at least it's a little less of an eyesore.

Lastly, in writing printf, I added a va_arg for pointers because when I
was casting a pointer to a uint64_t through va_arg, I got a bunch of
garbage in the prefix. This fixes that by considering every pointer as a
void* in va_arg and then casting it to an uint64_t.