Refactored linker script and boot to be more architecture agnostic.
Moved the linker script dependency to a dependency of i386. Added
--spawn_strategy=standalone to the qemu.sh script. Unfortunately, there
is a bug in bazel seen here,
https://github.com/bazelbuild/bazel/issues/716, that requires sandboxing
to be disabled to allow the linker script dependency to persist to the
main kernel dependency. It's not ideal but if they ever fix this bug,
the code as is should function with sandboxing.