Refactored boot to use a singleton mmap and provided local toolchain.
Boot wasn't that pretty so I modified it to use a singleton class for
the mmap entries and manipulation. Furthermore, I added a toolchain for
running as local in hopes that there will be a nice solution to ignoring
directories with //... based on arch in Bazel and thus everything is buildable
without ignoring specific rules.