Added dynamic memory allocator and enum -> enum class.
This had to happen at some point! This is a pretty inefficient memory
allocator that does not use paging but instead memory regions in an
arraylist. This should be replaced at some point but it seems to work
for now.

Furthermore, the allocator is not threadsafe (synchronization doesn't exist yet).

Adding a doc on the memory allocator.