Added //:qemu rule to run QEMU.
Removed qemu.sh since it was kinda messy and replaced it with a python
script qemu.py. Now, once the dependencies are installed (should
automate this as a dependency at some point using genrules and bazel
libraries), you can call `bazel run //:qemu --cpu=i386` to run the OS in
QEMU on the i386 architecture.