Adding StatusOr and Either along with tests.
It would be helpful to have a class for return values that may or may not return an error. In addition, we must utilize this StatusOr/Status class for errors because we cannot use C++ exceptions for errors. These require runtime support, which does not currently exist in this kernel. See https://wiki.osdev.org/C%2B%2B_Exception_Support for more why.