Adding a working dummy interrupt handler.
This commit fixes a bug where the limit and address were swapped in the IDTR value
for the Interrupt Descriptor Table (IDT). In addition, this commit moves tty
initialization to pre_kernel_main. Most importantly, it adds an interrupt handler
that prints out a string for interrupts on IRQ 0x80. This change does not mean
interrupts can be completely enabled (since we don't handle the recurring timer
interrupt or double faults).